Use State to trigger open & close
I've found that `handleProps` is called for each render of the parent container. This up calling `close` and thereby creating warnings in my console due to https://github.com/rackt/react-modal/blob/master/lib/helpers/focusManager.js#L36

What I've done is update the component to use its `state` to keep track of whether it's open or not. This fixes the warnings but may not be the most elegant solution to the problem.