[fixed] Make the modal portal render into body again
In c13fed9f3a1270a4ecfb6faf24355f2791948c00 behavior was changed
so that the modal portal ended up rendering into the app element.

In the default case where the appElement receives aria-hidden
when the modal is open, this meant that the modal itself was
also contained inside the aria-hidden element.  This limits the
ability for screenreaders to read the modal contents.

Server-side rendering should not be affected by this change
because componentDidMount is only fired on the client-side thus
having a reference to document.body there should be acceptable.

Upgrade Path:
  - Make sure that you are setting an app element (rather than
    using the default document.body) so that modal contents are
    not being aria-hidden