Update bundling scheme (#347)
* Update bundling scheme

This commit sets things up to work properly when
publishing to npm.  Prior to publishing the package
items from the src/ folder will be processed with Babel
and stored in the lib/ folder. The lib/ folder will
not be present in the repository.  The src/ folder similarly
will not be available when looking at the installed package.

In addition this commit also fixes #332 by adding the
`babel-plugin-transform-object-assign` plugin and removing
`lodash.assign`

* Update linting dependencies

This temporarily disables react/require-default-props
until it can be addressed in a separate commit.

* Make sure tests run using the proper source