[chore] changelog writer.
now we avoid the RCs and Alpha version, so they won't appear on the
log.

also, now we negate the log entries that are related to releases
using a filter, which is better than exclude the first entry
using Changelog::log_in_between_versions (that way, sometimes we
remove a valid entry).